Made different components for code
